Certificate download corrected
