+++ microsoft_student_partners/templates/certificates/project.html
@@ -110,7 +110,7 @@
                   </div><br><hr>
                   <div class="resister_text text-center">
                         <a target="_blank" href="{{ MEDIA_URL}}{{ context.img_file }}" class="boxed-btn-white">View Your Certificate</a>
-                        <a target="_blank" href="" class="boxed-btn-white" download="{{ MEDIA_URL }}{{ context.pdf_file }}">Download Certificate in pdf</a>
+                        <a target="_blank" href="{{ MEDIA_URL }}{{ context.pdf_file }}" class="boxed-btn-white" download>Download Certificate in pdf</a>
                         <a class="boxed-btn-white disable">Email (Disabled)</a>
                     </div>
                 </div>
